feat: single check_pr.py + full CI + fix BOM
